+++ src/modules/esm.mjs
@@ -1,0 +1,41 @@
+import path from 'path'; // const path = require('node:path');
+import { release, version } from 'os'; // const { release, version } = require('node:os');
+import { createServer as  createServerHttp } from 'http'; // const { createServer: createServerHttp } = require('node:http');
+import { fileURLToPath } from 'url';
+
+import './files/c.cjs'; // require('./files/c.cjs');
+
+const random = Math.random();
+
+const unknownObject = random > 0.5
+  ?
+    (await import('./files/a.json', { with: { type: 'json' } })).default
+  :
+    (await import('./files/b.json', { with: { type: 'json' } })).default;
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log('To terminate it, use Ctrl+C combination');
+});
+
+export {
+  unknownObject,
+  myServer,
+};
